+++ src/components/ThumbnailGenerator.tsx
@@ -4,6 +4,7 @@ import { Toolbar } from './Toolbar';
 import { PropertiesPanel } from './PropertiesPanel';
 import { StatusBar } from './StatusBar';
 import { LogoLibraryModal } from './LogoLibraryModal';
+import { MobileTextEditor } from './MobileTextEditor';
 import { ThumbnailCanvas } from './ThumbnailCanvas';
 import type { ActiveSnap } from '../types/snapping';
 
@@ -86,6 +87,7 @@ export const ThumbnailGenerator: React.FC<ThumbnailGeneratorProps> = ({ dragStat
       </div>
       <StatusBar />
       <LogoLibraryModal />
+      <MobileTextEditor />
     </div>
   );
 };